[refactoring][#66] interface 통일
- 기존에 state에 대한 interface와 data model에 대한 interface를 분리했었는데 사실상 같기 때문에 state에 대한 interface로 통일
- 이후 필요할 경우 data에 대한 interface 정의 예정
