+++ api/card.ts
@@ -1,6 +1,6 @@
 import { ICardList } from '@/interfaces/ICardList';
 import { TReturn } from '@/interfaces/api';
-import { ICardModel } from '@/interfaces/api/card';
+import { ICard } from '@/interfaces/ICard';
 import fetchApi from './fetch';
 import urls from './config/urls';
 
@@ -12,7 +12,7 @@ const CardApi = {
       data: { list },
     });
   },
-  addCard: (columnId: string, content: string): TReturn<ICardModel> => {
+  addCard: (columnId: string, content: string): TReturn<ICard> => {
     return fetchApi({
       url: urls.CARD,
       method: 'POST',
