+++ interfaces/IColumn.ts
@@ -5,3 +5,18 @@ export interface IColumn {
   title: string;
   items: ICardItems;
 }
+
+export interface addColumnDTO {
+  title: string;
+  items: ICardItems;
+}
+
+export interface dragColumnDTO {
+  targetColumnId: string;
+  dragColumnId: string;
+}
+
+export interface IDragColumnPayload {
+  targetColumnId: string;
+  dragColumnId: string;
+}
