--- interfaces/api/card-list.ts
@@ -1,3 +1,0 @@
-import { IColumnModel } from './column';
-
-export type ICardListModel = IColumnModel[];
