--- interfaces/api/column.ts
@@ -1,17 +1,0 @@
-import { ICardItmesModel } from './card';
-
-export interface IColumnModel {
-  id: string;
-  title: string;
-  items: ICardItmesModel;
-}
-
-export interface addColumnDTO {
-  title: string;
-  items: ICardItmesModel;
-}
-
-export interface dragColumnDTO {
-  targetColumnId: string;
-  dragColumnId: string;
-}
